[Core][Graphics] Support explicit descriptor binding indices in graphics backend
Added handling for explicit binding indices in descriptor bindings across Metal, Vulkan, and DirectX12 backends. Refactored logic to accommodate both implicit and explicit index specifications, ensuring proper indexing behavior. Updated data structures to use 32-bit integers for better compatibility and expanded descriptor configurations.

+++ Core/Src/Graphics/DirectX12/Dx12DescriptorSetManager.hpp
@@ -62,7 +62,7 @@ namespace KryneEngine
         struct LayoutData
         {
             eastl::array<ShaderVisibility, kRangeTypesCount> m_visibilities;
-            eastl::array<u16, kRangeTypesCount> m_totals;
+            eastl::array<u32, kRangeTypesCount> m_totals;
         };
 
         const LayoutData* GetDescriptorSetLayoutData(DescriptorSetLayoutHandle _layout);
